requirements.txt
watchdog
pyperclip
pystray
pillow

